2006-11-13 [colin]	2.6.0cvs42
	* src/plugins/pgpinline/pgpinline.c
	* src/plugins/pgpmime/pgpmime.c
		Very probably fix bug 1061, 'crash on
		opening signed email'. Another cast
		problem

+++ src/plugins/pgpinline/pgpinline.c
@@ -271,7 +271,7 @@ static gint pgpinline_check_signature(MimeInfo *mimeinfo)
 	gpgme_set_textmode(ctx, 1);
 	gpgme_set_armor(ctx, 1);
 	
-	gpgme_data_new_from_mem(&plain, textdata, strlen(textdata), 1);
+	gpgme_data_new_from_mem(&plain, textdata, (size_t)strlen(textdata), 1);
 	gpgme_data_new(&cipher);
 
 	data->sigstatus = sgpgme_verify_signature(ctx, plain, NULL, cipher);
@@ -398,7 +398,7 @@ static MimeInfo *pgpinline_decrypt(MimeInfo *mimeinfo)
 	}
 
 	debug_print("decrypting '%s'\n", textdata);
-	gpgme_data_new_from_mem(&cipher, textdata, strlen(textdata), 1);
+	gpgme_data_new_from_mem(&cipher, textdata, (size_t)strlen(textdata), 1);
 
 	plain = sgpgme_decrypt_verify(cipher, &sigstat, ctx);
 	if (sigstat && !sigstat->signatures)
@@ -519,7 +519,7 @@ static gboolean pgpinline_sign(MimeInfo *mimeinfo, PrefsAccount *account)
 	
 	fclose(fp);
 		
-	gpgme_data_new_from_mem(&gpgtext, textstr, strlen(textstr), 0);
+	gpgme_data_new_from_mem(&gpgtext, textstr, (size_t)strlen(textstr), 0);
 	gpgme_data_new(&gpgsig);
 	if ((err = gpgme_new(&ctx)) != GPG_ERR_NO_ERROR) {
 		debug_print(("Couldn't initialize GPG context, %s"), gpgme_strerror(err));
@@ -687,7 +687,7 @@ static gboolean pgpinline_encrypt(MimeInfo *mimeinfo, const gchar *encrypt_data)
 	fclose(fp);
 
 	/* encrypt data */
-	gpgme_data_new_from_mem(&gpgtext, textstr, strlen(textstr), 0);
+	gpgme_data_new_from_mem(&gpgtext, textstr, (size_t)strlen(textstr), 0);
 	gpgme_data_new(&gpgenc);
 	if ((err = gpgme_new(&ctx)) != GPG_ERR_NO_ERROR) {
 		debug_print(("Couldn't initialize GPG context, %s"), gpgme_strerror(err));
